feat: added login button to landing page, moved ipfs content to dashboard page
